feat(codex): add tmux mode for CLI transport
Add opt-in tools.codex.tmux_mode config parsing and validation, propagate the setting into codex runtime metadata, and wrap codex CLI command construction in a detached csa-<ULID> tmux session while preserving stdout/stderr/stdin pipe capture.

Split executor command construction and legacy transport modules to keep touched files under monolith gates.

AI Reviewer Metadata: GitNexus impact checked and detect-changes run. Verified with cargo clippy --workspace --all-features --jobs 1 -- -D warnings, cargo test -p csa-executor --jobs 1, cargo test -p cli-sub-agent --jobs 1, cargo fmt --check, and CARGO_BUILD_JOBS=1 just pre-commit.
