+++ crates/csa-executor/src/executor.rs
@@ -6,7 +6,6 @@ use csa_process::ExecutionResult;
 use csa_session::state::{MetaSessionState, ToolState};
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
-use std::ffi::OsString;
 use std::path::Path;
 use tokio::process::Command;
 
@@ -32,6 +31,10 @@ use arg_helpers::{
     effective_gemini_model_override, gemini_include_directories,
 };
 
+#[path = "executor_codex_tmux.rs"]
+mod codex_tmux;
+#[path = "executor_command.rs"]
+mod command;
 #[path = "executor_env.rs"]
 pub(crate) mod executor_env;
 #[path = "executor_pre_session.rs"]
@@ -268,59 +271,6 @@ impl Executor {
         }
     }
 
-    /// Inject environment variables from global config into a Command.
-    pub fn inject_env(cmd: &mut Command, env_vars: &HashMap<String, String>) {
-        for (key, value) in env_vars {
-            if !Self::STRIPPED_ENV_VARS.contains(&key.as_str()) {
-                cmd.env(key, value);
-            }
-        }
-    }
-
-    /// Build a configured Command ready for execution (without spawning).
-    pub fn build_command(
-        &self,
-        prompt: &str,
-        tool_state: Option<&ToolState>,
-        session: &MetaSessionState,
-        extra_env: Option<&HashMap<String, String>>,
-    ) -> (Command, Option<Vec<u8>>) {
-        // Prepend CSA identity preamble for claude-code (#1397).
-        let preamble_buf;
-        let prompt = match self {
-            Self::ClaudeCode { .. } => {
-                preamble_buf =
-                    format!("{}{prompt}", Self::csa_sub_agent_identity_preamble(session));
-                preamble_buf.as_str()
-            }
-            _ => prompt,
-        };
-        let mut cmd = self.build_base_command(session);
-        if matches!(self, Self::GeminiCli { .. } | Self::AntigravityCli { .. }) {
-            Self::strip_gemini_inherited_env(&mut cmd);
-        }
-        if let Some(env) = extra_env {
-            Self::inject_env(&mut cmd, env);
-        }
-        self.inject_csa_owned_env(&mut cmd, session);
-        executor_env::inject_git_guard_env(&mut cmd);
-        let gemini_include_directories =
-            gemini_include_directories(extra_env, prompt, Some(Path::new(&session.project_path)));
-        let (prompt_transport, stdin_data) = self.select_prompt_transport(prompt);
-        self.append_tool_args_with_transport(
-            &mut cmd,
-            prompt,
-            tool_state,
-            prompt_transport,
-            &gemini_include_directories,
-        );
-        if matches!(self, Self::Codex { .. }) {
-            sanitize_env_for_codex(&mut cmd);
-            cmd = Self::sanitize_codex_command_args(cmd);
-        }
-        (cmd, stdin_data)
-    }
-
     /// Execute a task with full session context.
     pub async fn execute(
         &self,
@@ -440,85 +390,6 @@ impl Executor {
         Ok(result)
     }
 
-    /// Build command for execute_in() legacy path.
-    pub(crate) fn build_execute_in_command(
-        &self,
-        prompt: &str,
-        work_dir: &Path,
-        extra_env: Option<&HashMap<String, String>>,
-    ) -> (Command, Option<Vec<u8>>) {
-        let mut cmd = Command::new(self.executable_name());
-        cmd.current_dir(work_dir);
-        // Strip recursive-invocation guard vars (same as build_base_command).
-        for var in Self::STRIPPED_ENV_VARS {
-            cmd.env_remove(var);
-        }
-        if matches!(self, Self::GeminiCli { .. } | Self::AntigravityCli { .. }) {
-            Self::strip_gemini_inherited_env(&mut cmd);
-        }
-        if let Some(env) = extra_env {
-            Self::inject_env(&mut cmd, env);
-        }
-        executor_env::inject_git_guard_env(&mut cmd);
-        let gemini_include_directories =
-            gemini_include_directories(extra_env, prompt, Some(work_dir));
-        self.append_yolo_args(&mut cmd);
-        self.append_model_args(&mut cmd);
-        if matches!(self, Self::GeminiCli { .. } | Self::AntigravityCli { .. }) {
-            append_gemini_include_directories_args(&mut cmd, &gemini_include_directories);
-        }
-        if matches!(self, Self::Codex { .. })
-            && let Some(env) = extra_env
-        {
-            cmd.args(codex_notify_suppression_args(env));
-        }
-        let (prompt_transport, stdin_data) = self.select_prompt_transport(prompt);
-        if matches!(prompt_transport, PromptTransport::Argv) {
-            self.append_prompt_args(&mut cmd, prompt);
-        } else {
-            self.append_prompt_args_with_transport(&mut cmd, prompt, prompt_transport);
-        }
-        if matches!(self, Self::Codex { .. }) {
-            sanitize_env_for_codex(&mut cmd);
-            cmd = Self::sanitize_codex_command_args(cmd);
-        }
-        (cmd, stdin_data)
-    }
-
-    fn sanitize_codex_command_args(cmd: Command) -> Command {
-        let program = cmd.as_std().get_program().to_os_string();
-        let current_dir = cmd.as_std().get_current_dir().map(|dir| dir.to_path_buf());
-        let envs = cmd
-            .as_std()
-            .get_envs()
-            .map(|(key, value)| (key.to_os_string(), value.map(|v| v.to_os_string())))
-            .collect::<Vec<_>>();
-        let mut args = cmd
-            .as_std()
-            .get_args()
-            .map(|arg| arg.to_string_lossy().into_owned())
-            .collect::<Vec<_>>();
-        sanitize_args_for_codex(&mut args);
-
-        let mut sanitized = Command::new(program);
-        if let Some(dir) = current_dir {
-            sanitized.current_dir(dir);
-        }
-        for (key, value) in envs {
-            match value {
-                Some(value) => {
-                    sanitized.env(key, value);
-                }
-                None => {
-                    sanitized.env_remove(key);
-                }
-            }
-        }
-        let os_args = args.into_iter().map(OsString::from).collect::<Vec<_>>();
-        sanitized.args(os_args);
-        sanitized
-    }
-
     /// Environment variables to strip from child processes.
     const STRIPPED_ENV_VARS: &[&str] = executor_env::STRIPPED_ENV_VARS;
 
@@ -555,10 +426,7 @@ impl Executor {
         cmd.env("CSA_PROJECT_ROOT", &session.project_path);
         Self::inject_session_path_env(cmd, session);
 
-        // CSA_TOOL: tells the child process which tool it is running as
         cmd.env("CSA_TOOL", self.tool_name());
-        // CSA_PARENT_TOOL: tells the child process which tool its parent is
-        // (read from current process's CSA_TOOL, set by the parent CSA instance)
         if let Ok(current_tool) = std::env::var("CSA_TOOL") {
             cmd.env("CSA_PARENT_TOOL", current_tool);
         }
